Add swiftUI dependencies to their own file
